Update go version in generated mod files to go1.21 (#169)
If Blueprint is being used with go1.22 then running go mod download as part of the build process to add a toolchain directive to the generated go.mod files. This toolchain directive is something that old versions of Go (<1.21) cannot parse which results in errors when building docker containers.
Upgrading the go version to 1.21 in the generated code is a simple fix for this issue.